+++ docs/CHANGELOG.md
@@ -18,6 +18,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   by switching to a two-step process: request a task and then check on the result.
   See [Asynchronous calls to jobs](./user/async-job-calls.md) guide for more details.
 
+### Changed
+- `pydantic` package has been upgraded to version 2 to prevent from conflicts with other up-to-date packages.
+  `pydantic` is one of the dependencies of the `racetrack-client` package.
+
 ## [2.26.0] - 2024-01-24
 ### Changed
 - Job type plugins can now build job images from a single Dockerfile.